+++ docs/source/conf.py
@@ -4,8 +4,8 @@ sys.path.insert(0, os.path.abspath("../.."))
 
 project = "scCS"
 author = "Emil Kriukov"
-version = "0.3.2"
-release = "0.3.2"
+version = "0.5.0"
+release = "0.5.0"
 copyright = "2026, Emil Kriukov"
 html_logo = "_static/logo.png"
 extensions = [